feat: move copy function to file operate component
